degree int->double

+++ C2H6O/app/src/main/java/com/example/team1288/c2h6o/ListViewAdapter.java
@@ -78,7 +78,7 @@ public class ListViewAdapter extends BaseAdapter {
     }
 
     // 아이템 데이터 추가를 위한 함수. 개발자가 원하는대로 작성 가능.
-    public void addItem(Drawable icon, int idNum, byte[] picture, String name, int degree, int price, String explain, Drawable arrow) {
+    public void addItem(Drawable icon, int idNum, byte[] picture, String name, double degree, int price, String explain, Drawable arrow) {
         ListViewItem item = new ListViewItem();
 
         item.setIcon(icon);
